Extract value parsing into parser
